[Builtin_cd] finished norme
