+++ src/builtins/builtin_cd.c
@@ -8,95 +8,7 @@
 #include "unistd_42.h"
 #include "ft_printf.h"
 
-static void	rule_5(t_string *curpath, bool *must_print_pwd, const char *dir)
-{
-	t_string	next;
-	const char	*cdpaths;
-
-	if (var_get("CDPATH", &cdpaths) == OK_)
-	{
-		string_init(&next);
-		while ((cdpaths = get_next_path(&next, cdpaths)) != NULL)
-		{
-			string_cat(&next, dir);
-			if (is_dir(next.str))
-			{
-				string_shutdown(curpath);
-				*curpath = next;
-				*must_print_pwd = cdpaths[0] != ':' && cdpaths[0] != '\0';
-				return ;
-			}
-		}
-		string_shutdown(&next);
-	}
-	string_cat(curpath, dir);
-}
-
-int		physical_resolution(t_string *curpath)
-{
-	char	*new_pwd;
-
-	if (chdir(curpath->str) == -1)
-	{
-		error_set_context("%s: %s", curpath->str, strerror(errno));
-		return (ERROR_);
-	}
-	new_pwd = getcwd(NULL, 0);
-	if (new_pwd == NULL)
-		return (ERROR_);
-	var_set("PWD", new_pwd);
-	free(new_pwd);
-	return (OK_);
-}
-
-int		logical_resolution(t_string *curpath, t_string backup,
-							const char *pwd)
-{
-	backup.len = 0;
-	var_get("PWD", &pwd);
-	if (curpath->str[0] != '/')
-	{
-		string_insert(curpath, 0, "/", 1);
-		string_insert(curpath, 0, pwd, ft_strlen(pwd));
-	}
-	if (builtin_cd_rule_8(curpath) == -1)
-		return (ERROR_);
-	if (!ft_strncmp(pwd, curpath->str, ft_strlen(pwd))
-		&& curpath->str[ft_strlen(pwd)] == '/')
-	{
-		string_clone(&backup, curpath);
-		string_remove(curpath, 0, ft_strlen(pwd));
-		string_insert(curpath, 0, ".", 1);
-	}
-	if (chdir(curpath->str) == -1)
-	{
-		error_set_context("%s: %s", curpath->str, strerror(errno));
-		return (ERROR_);
-	}
-	var_set("OLDPWD", pwd);
-	var_set("PWD", backup.len != 0 ? backup.str : curpath->str);
-	return (OK_);
-}
-
-void	get_base_path(t_string *curpath, bool *must_print_pwd, const char *dir)
-{
-	t_string	comp;
-
-	if (dir[0] == '/')
-		string_cat(curpath, dir);
-	else
-	{
-		string_init(&comp);
-		get_next_component(&comp, curpath->str);
-		if (ft_streq(comp.str, ".") || ft_streq(comp.str, ".."))
-			string_cat(curpath, dir);
-		else
-			rule_5(curpath, must_print_pwd, dir);
-		string_shutdown(&comp);
-	}
-}
-
-const char	*get_dir(bool *must_print_pwd, const char *arg)
+static const char	*get_dir(bool *must_print_pwd, const char *arg)
 {
 	const char	*dir;
 	const char	*home;
@@ -124,7 +36,8 @@ const char	*get_dir(bool *must_print_pwd, const char *arg)
 	return (dir);
 }
 
-int		get_options(bool *p, const char **arg, int ac, const char *const *av)
+int					get_options(bool *p, const char **arg, int ac,
+								const char *const *av)
 {
 	t_opt	opt;
 	char	ret;
@@ -152,7 +65,7 @@ int		get_options(bool *p, const char **arg, int ac, const char *const *av)
 	return (OK_);
 }
 
-int		cd(int ac, const char *const *av)
+int					cd(int ac, const char *const *av)
 {
 	t_string	curpath;
 	int			ret;
@@ -178,7 +91,7 @@ int		cd(int ac, const char *const *av)
 	return (ret);
 }
 
-int		builtin_cd(int ac, const char *const *av)
+int					builtin_cd(int ac, const char *const *av)
 {
 	if (cd(ac, av) == OK_)
 		return (0);
